ACD-36 Added support for non-OSGi VM.
Replace bundles with domains wherever it was more generic.
Added class or instance location in debug output to better troubleshoot cases where the code is in a base class of a given object.
Added system property expansion for domain locations and file permission names and made it configurable.
Fixed when using backdoor to get bundle to rely on the result even if null
Fixed printout