ACD-40 Fixed to properly handle fragment bundles.
Added support for getting a domain from a class in the backdoor.
Enhanced to report the canonicalized path for file permissions.
Modified to still put a break when somthing is calling doPrivileged() on behalf of someone else as we cannot assume the context will be the same as the stack prior to the call. Did at least move the stack break to be after the caller of those special places.